+++ THManager/THManager/Models/AccountTradeHistory.cs
@@ -11,8 +11,8 @@ namespace THManager.Models
         public DateTime Time { get; set; }
         public string Side { get; set; }
         public string Pair { get; set; }
-        public decimal Price { get; set; }
-        public decimal Quantity { get; set; }
+        public double Price { get; set; }
+        public double Quantity { get; set; }
         public decimal DollarQuantity { get; set; }
         public double BalanceUSDT { get; set; }
         public double Profit { get; set; }
